Remove rabbitMq from catalog

+++ fiscalization-app-services/catalog-management/src/main/java/hr/unizg/fer/sudec/karlo/catalogManagement/service/CatalogItemRepository.java
@@ -3,5 +3,5 @@ package hr.unizg.fer.sudec.karlo.catalogManagement.service;
 import hr.unizg.fer.sudec.karlo.catalogManagement.entity.CatalogItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CatalogItemRepository extends JpaRepository<CatalogItem, Integer> {
+public interface CatalogItemRepository extends JpaRepository<CatalogItem, Long> {
 }
